==> 00_collection.ipynb <==
# NLP를 활용한 레시피 추천 챗봇 구현
**개요**
- Collection 
- Preprocessing
- EDA
- Embeddings
- Modeling
- Streamlit

## 1. Collection (웹 크롤링 : 스크래핑)
- 수집할 데이터 : 요리이름, 카테고리, 난이도, 소요시간, 재료, 요리방법, 요리설명, 링크, 이미지링크
- 수집 대상 사이트 : 우리의 식탁(https://wtable.co.kr/recipes)
- 수집 방법 : 웹 크롤링(스크래핑)
    - BeautifulSoup
    - requests

### 라이브러리 import

# 데이터 분석
import numpy as np
import pandas as pd

# 웹 스크래핑
import requests
from bs4 import BeautifulSoup as bs

# HTTP 클라이언트
import urllib3
urllib3.disable_warnings()

# 진행시간 표시
import time
from tqdm.notebook import tqdm
tqdm.pandas()

### 웹 스크래핑 진행하기

def get_link(token):
    "음식 상세정보 링크 반환"
    link = f'https://wtable.co.kr/recipes/{token}?location=recipe_home'
    return link

# 카테고리 코드
category = {'메인요리': 1,
            '초대요리': 2,
            '간식': 5,
            '간단요리': 126,
            '밑반찬': 183,
            '채식': 184,
            '한식': 185,
            '양식': 186,
            '일식': 187,
            '중식': 188,
            '퓨전': 189,
            '분식': 190,
            '안주': 191,
            '베이킹': 192,
            '다이어트': 193,
            '도시락': 214,
            '키토': 242,
            '오븐 요리': 273
            }

def get_one_page(cuisine, start):
    "cuisine의 start번째부터 40개 수집"
    url_base = f'https://wtable.net/api_v2/theme/recipe/list?&platform=web&'
    url_base2 = f'theme_id={category[cuisine]}&offset={start}&limit=40'
    url = url_base + url_base2

    response = requests.get(url, verify=False)
    data = response.json()['data']

    level_list = list(map(lambda x: x['level'], data))
    time_list = list(map(lambda x: x['time'], data))
    title_list = list(map(lambda x: x['title'], data))
    link_list = [get_link(i['token']) for i in data]
    img_list = list(map(lambda x: x['profile_img'], data))


    df = pd.DataFrame({'요리':title_list,                 
                       '종류':cuisine,
                       '난이도':level_list,
                       '소요시간':time_list,
                       '링크':link_list,
                       '사진':img_list})

    return df

def get_all_pages(cuisine):
    "모든 페이지 수집"
    start = 0
    page_list = []
    while True:
        df_one_page = get_one_page(cuisine, start)
        if df_one_page.empty:
            break
        else:
            page_list.append(df_one_page)
            time.sleep(0.01)
            start += 40

    df_all_page = pd.concat(page_list, ignore_index=True)
    
    return df_all_page

# 모든 카테고리를 대상으로 get_all_pages 적용하기
cuisines = []
for i in category:
    cuisines.append(get_all_pages(i))
result = pd.concat(cuisines, ignore_index=True)

def get_details(url):
    "링크에서 재료와 요리방법, 설명 수집 함수"
    try: 
        response = requests.get(url, verify=False)
        soup = bs(response.text)    
        tmp = soup.find_all('div', {'class':'Text__Description02-sc-1qy6bx2-0 fCbbYE'})
        ingredients = [i.find_all('div')[0].text for i in tmp]
        tmp2 = soup.find_all('p', {'class':'Text__Pre01-sc-1qy6bx2-2 enJPxd'})
        cook = [i.text for i in tmp2]
        description = soup.find('p', {'class':'RecipeDetailstyle__Description-q7sykd-14 IdQIJ'}).text
        return ingredients, cook, description
    
    except:
        return None, None, None

# 링크에서 재료와 요리방법, 설명 가져오기
details = result['링크'].progress_map(get_details)

result['재료'] = [i[0] for i in details]
result['요리방법'] = [i[1] for i in details]
result['설명'] = [i[2] for i in details]

### 수집한 데이터 확인하기

# 수집한 데이터 구성 요약
result.info()

# 요리, 종류, 설명만 가져와서 보기
result[['요리', '종류', '설명']]

# 데이터 미리보기
result.head(1)

### 수집한 데이터 저장하기

result.to_csv('data/raw_recipes.csv', index=False)

raw_recipes = pd.read_csv('data/raw_recipes.csv')
raw_recipes.head()
