==> 01_preprocessing.ipynb <==
# NLP를 활용한 레시피 추천 챗봇 구현
**개요**
- Collection 
- Preprocessing
- EDA
- Embeddings
- Modeling
- Streamlit

## 2. Preprocessing (전처리)
- 수집한 데이터 : 요리이름, 카테고리, 난이도, 소요시간, 재료, 요리방법, 요리설명, 링크, 이미지링크
- 전처리할 데이터
    - 재료에 섞인 불용어 위주로 제거
    - 설명에 '비만', '당뇨'가 들어가는 음식 체크


### 라이브러리 import

# 데이터 분석
import numpy as np
import pandas as pd

# 진행시간 표시
import time
from tqdm.notebook import tqdm
tqdm.pandas()

# 정규표현식
import re

### 데이터 불러오기

# 크롤링한 데이터 불러오기
raw_recipes = pd.read_csv('data/raw_recipes.csv')

raw_recipes.shape

raw_recipes.head(1)

### 전처리 진행하기: 재료

# 수기로 발견한 대표적인 불용어와 많은 레시피에 두루 포함되는 조미료 계열을 리스트 형태로 변수 할당
stop_words = ['송송','썬','삶은','짤주머니','사각','X','곱게','리덕션','채썬',
              '냉장','늙은','판','미니','집','흰','두절','레드','모양','같이','잘','익은',
              '고운','익힌','주머니','건세모가사리','빨강','썰은','소분한','작은',
              '마른','벗긴','밥심','뜨거운','따듯한','씨를','크기로','제거','생물','빨간',
              '갈은','지은','데친','국물용','부순','꼬치','찬','차가운','빨간','중하',
              '컬러','구멍','따뜻한','A','틀','초밥용','몽글몽귤','튜브','두꺼운','크기','모둠',
              '해감한','와','구운','굵은','그린','굵게','익은','시판','것','없는','cm','상온',
              '삶은','무첨가','이지잇','잘게', '자른','깐','찐','씻은','잡채용',
              '고운', '면수']

# 텍스트 전처리 (1차)
def preprocessing(text):
    '''
    문자열을 전달하면 불용어, 중복, 공백을 제거하여 리스트 형태로 반환하는 함수

    Args:
    - 문자열 (예 : "['순살 고등어', '청주(또는 맛술)', '소금', '소금']")

    Returns:
    - 리스트 (예 : ['순살 고등어', '청주', '소금'])

    Example:
    >>> preprocessing("['순살 고등어', '청주(또는 맛술)', '소금', '소금']")
    ['순살고등어', '청주', '소금']
    '''
    process = eval(text)

    process = [elem for elem in process if not re.search(r'\*', elem)]
    process = [elem.replace('\xa0', '') for elem in process]
    process = [elem.replace(' ', '') for elem in process]

    process = ' '.join(set(process))
    process = re.sub(r'\([^)]*\)', '', process).strip()
    process = re.sub(r'[^가-힣ㄱ-ㅎㅏ-ㅣa-zA-Z]', ' ', process).strip()
    pattern = '|'.join(stop_words)
    process = re.sub(pattern, '', process)
    process = re.sub(r'또는', ' ', process)

    process = process.split(' ')

    process = [elem for elem in process if elem != '']
    
    return process

# 데이터 프레임 복사
df = raw_recipes.copy()

# 재료 컬럼에 텍스트 전처리 (1차) 적용
df['재료'] = df['재료'].progress_map(preprocessing)
df['재료']

# 확인하기
df.head(2)

# 재료수 파생변수를 생성하는 함수
def ingredients_count(text):
    ing_count = len(text)
    return ing_count

# 재료 컬럼에 재료수 파생변수 생성
df['재료수'] = df['재료'].progress_map(ingredients_count)
df.head(2)

# 텍스트 전처리 (2차)
def preprocessing2(list):
    '''
    리스트를 전달하면 공백으로 구분된 하나의 문자열로 반환하는 함수

    Args:
    - 리스트 (예 : ['순살 고등어', '청주', '소금'])

    Returns:
    - 문자열 (예 : "순살고등어 청주 소금")

    Example:
    >>> preprocessing2(['순살 고등어', '청주', '소금'])
    "순살고등어 청주 소금"
    '''
    process = str(list)
    process = eval(process)
    process = ' '.join(set(process))

    return process

# 재료 컬럼에 텍스트 전처리 (2차) 적용
df['재료'] = df['재료'].progress_map(preprocessing2)
df.head(2)

### 전처리 진행하기: 설명

# 불용어(조사) 리스트
stopwords = ['는', '은', '이', '에', '의', '가', '이', '도','을', '를', '으로', '에서', '와', '과', '한', '하고', '을', '로', '에', '나', '다가', '하게', '해서', '들', '고', '도', '에도']

def preprocess_description(description, stopwords):
    # 한글만 남기기
    words = re.sub(r'[^가-힣]', ' ', description)
    # 설명을 띄어쓰기 단위로 나누기
    words = words.split()
    # 각 단어에서 맨 뒤에 붙어있는 불용어 제거
    cleaned_words = []
    for word in words:
        for stopword in stopwords:
            if word.endswith(stopword):
                word = word[:-len(stopword)]
        cleaned_words.append(word)
    
    return cleaned_words

# 전처리된 설명 컬럼 추가
df['전처리된_설명'] = df['설명'].apply(lambda x: preprocess_description(x, stopwords))

df.head(5)

# 질병 유무 판단
def extract_disease(words):
    # 질병 유무 판단
    if any(keyword in words for keyword in ['비만', '다이어트', '비만예방']):
        return '비만'
    elif any(keyword in words for keyword in ['당뇨', '당뇨병', '혈당']):
        return '당뇨'
    else:
        return '일반'

df['질병'] = df['전처리된_설명'].apply(extract_disease)

df.head(5)

# '전처리된_설명' 컬럼 제거
df.drop(columns=['전처리된_설명'], inplace=True)

df.head(10)

df[df['질병'] == '비만']

# csv 파일로 저장하고 불러오기
df.to_csv('data/preprocessed_recipes.csv', index=False)
n_df = pd.read_csv('data/preprocessed_recipes.csv')

# 불용어 제거 확인하기 
filtered_df = n_df[n_df['재료'].apply(lambda x: any(word in x for word in stop_words))]
filtered_df 

# 텍스트 전처리 전과 후를 비교하는 함수
def sample_test(origin_df, processed_df, column_name):
    sample = origin_df.sample(1)[column_name]
    processed_sample = processed_df.iloc[list(sample.index)][column_name]
    with pd.option_context('display.max_colwidth', None):
        return display(sample.to_string(index=False), processed_sample.to_string(index=False))

# 실행할 때 마다 sample 무작위로 변경하면서 확인 가능
sample_test(raw_recipes, df, '재료')
