==> 03_embeddings.ipynb <==
# NLP를 활용한 레시피 추천 챗봇 구현
**개요**
- Collection 
- Preprocessing
- EDA
- Embeddings
- Modeling
- Streamlit

## 4. Embeddings
**모델링 과정**
1. 사전 학습된 모델을 Sentence Transformer 를 통해 불러오기
2. 수집하고 기본적인 전처리를 거친 데이터를 위의 모델을 통해 임베딩 벡터로 변환한 파생변수 생성하기
3. input 으로 사용자가 재료를 포함한 문자열을 입력하면 문장을 벡터화하여 기존의 임베딩 벡터와 코사인 유사도를 구하는 함수
4. 코사인 유사도 상위 기준으로 n개를 output 으로 추천
5. OpenAI 의 API 를 활용하여 사용자의 문장 형태 input 의 의도를 파악하여 모델링 함수를 실행하도록 연동
6. streamlit 에 연동하여 배포


### 라이브러리 import

# 데이터 분석
import pandas as pd
import numpy as np

# 데이터 시각화
import matplotlib.pyplot as plt
import seaborn as sns

# 진행시간 표시
import swifter
from tqdm.notebook import tqdm
tqdm.pandas()

# 파이토치
import torch

# 문장 임베딩, transformer 유틸리티
from transformers import AutoTokenizer, AutoModel
from sentence_transformers import SentenceTransformer, util
from sentence_transformers import SentenceTransformer, models

# 객체 복사
import copy

# JSON 형식 데이터 처리
import json

# 데이터 수집
import requests

# 데이터베이스 활용
import sqlite3 
import pickle

# OpenAI API 활용
import openai 
import os # 운영체제
import sys # 파이썬 변수, 함수 엑세스 
from dotenv import load_dotenv # 환경 변수 로드(API Key 보안)

os.environ["TOKENIZERS_PARALLELISM"] = "false"
load_dotenv()    
openai.api_key = os.getenv('OPENAI_API_KEY')

# 실행 os 확인
cur_os = sys.platform
cur_os

### 데이터 불러오기

df = pd.read_csv('data/preprocessed_recipes.csv')
df.shape

df.head(1)

difficulty_map = {1: '쉬움', 2: '보통', 3: '어려움'}
df['난이도'] = df['난이도'].map(difficulty_map)

df.info()

### 전이학습을 통한 임베딩 벡터 생성
- 사전학습 모델을 활용
- feature 생성 및 비교

### feature 생성

df['feature'] = df['재료'] + df['요리'] + df['설명'] + df['종류'] + df['난이도'] + df['요리방법'] + df['질병']
df['feature']

df.to_csv('data/feature_recipes.csv', index=False)

### 사전 학습 모델
- model : Sentence Transformer
- pre trained model : `jhgan/ko-sroberta-multitask`

model_name = 'jhgan/ko-sroberta-multitask'

model = SentenceTransformer(model_name)
model

# feature1
feature_name = 'feature'
df[f'{model_name}-' + f'{feature_name}'] = df['feature'].progress_apply(lambda x: model.encode(x))

len(df.columns)

df.columns = ['요리', '종류', '난이도', '소요시간', '링크', '사진', '재료', '요리방법', '설명', '재료수', '질병', 'feature', 'ko-sroberta-multitask-feature']

df = df[['요리', '종류', '난이도', '재료', '요리방법', '설명', '재료수', '소요시간', '링크', '사진', '질병',
         'feature', 'ko-sroberta-multitask-feature']]
df

df.to_pickle('data/compact_kosroberta_recipes.pkl')
df = pd.read_pickle('data/compact_kosroberta_recipes.pkl')

df.head(1)

## DB에 임베딩 벡터 저장하고, 불러오는 과정
- 위에서 적용한 pickle 형식으로 데이터를 저장하는 것이 더 효율적이라고 판단됨

# # 데이터 프레임의 embedding
# def save_embeddings_to_db(df, db_file, model):
#     conn = sqlite3.connect(db_file)
#     cursor = conn.cursor()
#     cursor.execute('''CREATE TABLE IF NOT EXISTS embedding_table (id INTEGER PRIMARY KEY, embedding_data BLOB)''')
#     conn.commit()

#     for index, row in df.iterrows():
#         text = row['feature']
#         embedding = model.encode(text)
#         embedding_binary = pickle.dumps(embedding)
#         cursor.execute('''INSERT INTO embedding_table (id, embedding_data) VALUES (?,?)''', (index, embedding_binary))
#         conn.commit()

#     conn.close()

# save_embeddings_to_db(df, 'embedding.db', model)

# def add_embeddings_to_df(df, db_file):
#     conn = sqlite3.connect(db_file)
#     cursor = conn.cursor()

#     # 데이터베이스에서 임베딩 정보 불러오기
#     cursor.execute('''SELECT id, embedding_data FROM embedding_table''')
#     rows = cursor.fetchall()
    
#     # 불러온 임베딩 정보를 딕셔너리로 변환
#     embeddings_dict = {}
#     for row in rows:
#         index = row[0]
#         embedding_binary = row[1]
#         embedding = pickle.loads(embedding_binary)
#         embeddings_dict[index] = embedding
    
#     # 데이터프레임에 임베딩 정보 추가
#     embeddings = []
#     for index, row in df.iterrows():
#         if index in embeddings_dict:
#             embedding = embeddings_dict[index]
#             embeddings.append(embedding)
#         else:
#             embeddings.append(None)
#     df['embeddings'] = embeddings
    
#     conn.close()
#     return df


# add_embeddings_to_df(df, 'embedding.db')
# df
