==> 04_modeling.ipynb <==
# NLP를 활용한 레시피 추천 챗봇 구현
**개요**
- Collection 
- Preprocessing
- EDA
- Embeddings
- Modeling
- Streamlit

## 5. Modeling
**모델링 과정**
1. 사전 학습된 모델을 Sentence Transformer 를 통해 불러오기
2. 수집하고 기본적인 전처리를 거친 데이터를 위의 모델을 통해 임베딩 벡터로 변환한 파생변수 생성하기
3. input 으로 사용자가 재료를 포함한 문자열을 입력하면 문장을 벡터화하여 기존의 임베딩 벡터와 코사인 유사도를 구하는 함수
4. 코사인 유사도 상위 기준으로 n개를 output 으로 추천
5. OpenAI 의 API 를 활용하여 사용자의 문장 형태 input 의 의도를 파악하여 모델링 함수를 실행하도록 연동
6. streamlit 에 연동하여 배포


### 라이브러리 import

# 데이터 분석
import pandas as pd
import numpy as np

# 데이터 시각화
import matplotlib.pyplot as plt
import seaborn as sns

# 진행시간 표시
import swifter
from tqdm.notebook import tqdm
tqdm.pandas()

# 파이토치
import torch

# 문장 임베딩, transformer 유틸리티
from transformers import AutoTokenizer, AutoModel
from sentence_transformers import SentenceTransformer, util
from sentence_transformers import SentenceTransformer, models

# 객체 복사
import copy

# JSON 형식 데이터 처리
import json

# 데이터 수집
import requests

# 데이터베이스 활용
import sqlite3 
import pickle

# OpenAI API 활용
import openai 
import os # 운영체제
import sys # 파이썬 변수, 함수 엑세스 
from dotenv import load_dotenv # 환경 변수 로드(API Key 보안)

os.environ["TOKENIZERS_PARALLELISM"] = "false"
load_dotenv()
openai.api_key = os.getenv('OPENAI_API_KEY')

# 실행 os 확인
cur_os = sys.platform
cur_os

### 데이터 불러오기

df = pd.read_pickle('data/compact_kosroberta_recipes.pkl')
df.shape

### 파생 변수
- feature1 = '재료'
- feature2 = '재료' + '요리'
- feature3 = '재료' + '요리' + '종류'
- feature4 = '재료' + '요리' + '종류' + '난이도'
- feature5 = '재료' + '요리' + '종류' + '난이도' + '요리방법'
- **feature = '재료' + '요리' + '설명' + '종류' + '난이도' + '요리방법'**

### 모델 불러오기
- model : Sentence Transformer
- pre trained model : `jhgan/ko-sroberta-multitask`

model_name = 'jhgan/ko-sroberta-multitask'

model = SentenceTransformer(model_name)
model

df['feature'][0]

## 콘텐츠 추천 함수(코사인 유사도 기반)

def get_similar_recipes(query, model, df, exclude_ingredients=None, health=None, intent=None):
    data = {'요리': df['요리'], 'feature': df['ko-sroberta-multitask-feature']}
    query_encode = model.encode(query)
    query_encode = np.array(query_encode)  # NumPy 배열로 변환
    cos_scores = util.pytorch_cos_sim(query_encode, data['feature'])[0]
    top_results = torch.topk(cos_scores, k=100)

    top_indices = top_results[1].numpy()

    # exclude_ingredients 조건 적용
    if exclude_ingredients:
        if len(exclude_ingredients) > 1:
            exclude_mask = np.array([any(exclude in row for exclude in exclude_ingredients) for row in df['재료']])
        else:
            exclude_mask = np.array([exclude_ingredients[0] in row for row in df['재료']])
        exclude_idx = np.where(exclude_mask)[0]
        top_indices = np.setdiff1d(top_indices, exclude_idx)

    # health 조건 적용
    if health:
        include_mask = np.array([any(h in row for h in health) for row in df['질병']])
        include_idx = np.where(include_mask)[0]
        top_indices = np.intersect1d(top_indices, include_idx)

    filtered_scores = cos_scores[top_indices]
    top_filtered_results = torch.topk(torch.tensor(filtered_scores).clone().detach(), k=len(filtered_scores))

    if intent == '1':
        result_df = df.iloc[top_indices[top_filtered_results.indices], :][['요리', '종류', '재료', '설명', '난이도', '링크']]
        result_df = result_df.drop_duplicates(subset=['링크']).head(3)
    elif intent == '2':
        result_df = df.iloc[top_indices[top_filtered_results.indices], :][['요리', '종류', '재료', '설명', '난이도', '링크']]
        result_df = result_df.drop_duplicates(subset=['링크']).sample(3)
    else:
        result_df = df.iloc[top_indices[top_filtered_results.indices], :][['요리', '종류', '재료', '설명', '난이도', '링크']]
        result_df = result_df.drop_duplicates(subset=['링크']).head(5)
    
    return result_df

print(df.columns)

# 일반 추천
include = '여름 메뉴 추천해줘'
intent = '1'
exclude = []
health = []
result = get_similar_recipes(include, model, df, exclude_ingredients=exclude, health=health, intent=intent)
result

# 재추천
include = '여름 메뉴 추천해줘'
intent = '2'
exclude = []
health = []
result = get_similar_recipes(include, model, df, exclude_ingredients=exclude, health=health, intent=intent)
result

# 단일 제외
include = '여름 메뉴 추천해줘'
intent = '1'
exclude = ['오이']
health = []
result = get_similar_recipes(include, model, df, exclude_ingredients=exclude, health=health, intent=intent)
result

# 다중 제외
include = '여름 메뉴 추천해줘'
intent = '1'
exclude = ['오이', '토마토']
health = []
result = get_similar_recipes(include, model, df, exclude_ingredients=exclude, health=health, intent=intent)
result

# 건강 상태
include = '샐러드 레시피 알려줘'
intent = 1
exclude = []
health = ['비만']
result = get_similar_recipes(include, model, df, exclude_ingredients=exclude, health=health, intent=intent)
result

## ChatGPT 연동하기
### 세 가지 Role
- user : 마치 채팅하는 것처럼 ChatGPT에 직접 무언가를 물어보고 싶을 때 사용하는 role (ex. "Please explain what AI is")
- system : 유저에게 메시지를 받기 전에 모델을 초기화하거나 구성하려는 경우 사용하는 role (ex. "You are a helpful kindergarten teacher talking to children")
- assistant : 이전에 ChatGPT가 유저에게 보낸 메시지가 무엇인지 알려주는 role. 유저와 어시스턴트 사이의 대화를 저장하고 어시스턴트에게 이전 대화를 전달하여 응답값을 조정할 수 있음. (ex. 이전까지의 대화를 기억하게 함으로써 명사 -> 대명사로 이어지는 맥락을 이해할 수 있도록 해줌)
### Fine tuning
Role 을 지정하는 것 이외에 Fine-tuning 을 활용하는 것이 실제 서비스에서는 더 이상적인 형태
사용자의 데이터를 가지고 특정 태스크에 알맞게 커스텀하는 것
- `{"prompt" : "I really like this recipe!", "completion" : "positive"}`
- `{"prompt" : "I'd like to leave out the cucumber.", "completion" : "negative"}`
